+++ lib/testing/gpuTrackingBeamformSimulate.cpp
@@ -302,7 +302,7 @@ void gpuTrackingBeamformSimulate::main_thread() {
 
         // Unpack input data
         int dest_idx = 0;
-        for (int i = 0; i < input_buf->frame_size; ++i) {
+        for (size_t i = 0; i < input_buf->frame_size; ++i) {
             input_unpacked[dest_idx++] = HI_NIBBLE(input[i]) - 8;
             input_unpacked[dest_idx++] = LO_NIBBLE(input[i]) - 8;
         }
